Add a picking mesh to structures and units
