+++ emergence_lib/src/asset_management/mod.rs
@@ -13,7 +13,13 @@
 use std::any::TypeId;
 
 use self::{structures::StructureHandles, terrain::TerrainHandles, units::UnitHandles};
-use bevy::{asset::LoadState, prelude::*, utils::HashSet};
+use bevy::{
+    asset::LoadState,
+    prelude::*,
+    render::{mesh::Indices, render_resource::PrimitiveTopology},
+    utils::HashSet,
+};
+use hexx::{Hex, HexLayout, MeshInfo};
 
 pub(crate) mod manifest;
 pub(crate) mod structures;
@@ -121,4 +127,15 @@ impl AssetCollectionExt for App {
 
         self
     }
+}
+
+/// Constructs the mesh for a single hexagonal column with the specified height.
+fn hexagonal_column(hex_layout: &HexLayout, hex_height: f32) -> Mesh {
+    let mesh_info = MeshInfo::partial_hexagonal_column(hex_layout, Hex::ZERO, hex_height);
+    let mut mesh = Mesh::new(PrimitiveTopology::TriangleList);
+    mesh.insert_attribute(Mesh::ATTRIBUTE_POSITION, mesh_info.vertices.to_vec());
+    mesh.insert_attribute(Mesh::ATTRIBUTE_NORMAL, mesh_info.normals.to_vec());
+    mesh.insert_attribute(Mesh::ATTRIBUTE_UV_0, mesh_info.uvs.to_vec());
+    mesh.set_indices(Some(Indices::U16(mesh_info.indices)));
+    mesh
 }
