+++ emergence_lib/src/player_interaction/details.rs
@@ -120,7 +120,7 @@ fn get_details(
         SelectionType::None => SelectionDetails::None,
     };
 
-    info!("{selection_details:?}")
+    info!("Selection details: {selection_details:?}")
 }
 
 mod structure {
